--- src/main/java/http/response/HttpException.java
@@ -1,7 +1,0 @@
-package http.response;
-
-public class HttpException extends RuntimeException {
-	public HttpException(String message) {
-		super(message);
-	}
-}
