+++ src/main/java/http/response/HttpResponse.java
@@ -5,22 +5,26 @@ import java.io.File;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.Files;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import db.DataBase;
+import model.User;
+
 public class HttpResponse {
 	private static final Logger log = LoggerFactory.getLogger(HttpResponse.class);
 	private static final String NEWLINE = "\r\n";
 	private Map<String, String> headers;
-	private boolean hasBody;
 	private byte[] body;
 	private DataOutputStream dos;
 
 	public HttpResponse(OutputStream out) {
-		headers = new HashMap<>();
+		this.headers = new HashMap<>();
 		this.dos = new DataOutputStream(out);
 	}
 	
@@ -32,18 +36,36 @@ public class HttpResponse {
 	
 	public void forward(String path) {
 		headers.put("requestFirstLine", "HTTP/1.1 200 OK \r\n");
+		updateBody(path);
+		writeResponse();
+	}
+
+	private void updateBody(String path) {
 		try {
 			this.body = Files.readAllBytes(new File("./webapp" + path).toPath());
-			this.hasBody = true;
-			headers.put("Content-Length", Integer.toString(body.length));
 		} catch (IOException e) {
 			log.error("body를 읽어오면서 에러가 났다!!");
 		}
-		writeResponse();
+		if( path.contains("list.html")) {
+			this.body = addUserList(body);
+		}
+		headers.put("Content-Length", Integer.toString(body.length));
+	}
+	
+	private byte[] addUserList(byte[] body) {
+		StringBuilder sb = new StringBuilder();
+		sb.append(new String(body));
+		int offset = sb.indexOf("user-list");
+		List<User> users = new ArrayList<>(DataBase.findAll());
+		for (User user : users) {
+			sb.insert(offset + 14, "<tr>\r\n<th>#</th> <th>" + user.getUserId() + "</th> <th>" + user.getName()
+					+ "</th> <th>" + user.getEmail() + "</th><th></th>\r\n</tr>");
+		}
+		return sb.toString().getBytes();
 	}
 	
 	public boolean hasBody() {
-		return hasBody;
+		return body != null;
 	}
 	
 	public void setContentType(String type) {
@@ -65,7 +87,7 @@ public class HttpResponse {
 		}
 	}
 
-	protected void addHeader(String key, String value) {
+	public void addHeader(String key, String value) {
 		headers.put(key, value);
 	}
