+++ src/main/java/http/response/HttpResponseException.java
@@ -1,0 +1,7 @@
+package http.response;
+
+public class HttpResponseException extends RuntimeException {
+	public HttpResponseException(String message) {
+		super(message);
+	}
+}
